Move out metrics into the controller folder
